build: add clang-format and clang-tidy configs, fix high-signal tidy findings
Both configs are reverse-engineered from the codebase rather than adopted
wholesale.

clang-format describes the style already in use (4-space, attached braces,
Type* ptr, short if/case/function bodies on one line). Existing code is
deliberately NOT mass-reformatted: even this tuned config rewrites ~17.5k
lines across 218 files, and tuning plateaus there because the residue is
intentional hand-alignment (trailing-comment columns, hand-broken boolean
expressions). LLVM defaults would have rewritten ~153k lines. Format
incrementally with git-clang-format instead; vendored code is excluded via
.clang-format-ignore.

clang-tidy uses a narrow, high-signal check set. The broad misc-* checks
are off — they contribute 3.4k of the 3.7k findings (const-correctness
alone is 2.2k) and bury the checks that actually catch defects. Two
exclusions worth calling out: bugprone-switch-missing-default-case fights
the -Wswitch enabled in the previous commit (a default-less switch over an
enum is what surfaces newly-added enumerators), and performance-no-int-to-ptr
is inherent to the slab allocator's generational references.

Findings fixed (3709 -> 30):

- transport.cpp parsed Content-Length with atoi, which has no error
  channel: a malformed header parsed as 0 and produced a silent empty
  message instead of a framing error. Now strtol with range validation.
- json.cpp had an if/else with identical bodies; write_pos advances on
  every emitted byte, so the branch (and has_escape) was dead.
- interpreter.cpp multiplied the list index in u32 before widening to a
  pointer offset, in two places.
- indexer.cpp had 5 no-op std::move calls on trivially-copyable stubs.
- file.cpp used rewind(), which reports no error; now fseek(SEEK_SET).
- ir_optimize.cpp had a redundant cast to the sub-expression's own type.

Two findings are documented as intentional rather than changed: the
memcpy over rx::String is a deliberate trivially-relocatable idiom
(NOLINT with rationale), and the pre_if use-after-move is a false
positive — the two moves are guarded by mutually exclusive conditions.

The 30 remaining are mostly clang-analyzer null-path warnings in
semantic.cpp, plausibly infeasible given the never-null error_type
sentinel design, but left as a triage backlog rather than suppressed
unverified.

Build stays warning-clean; full suite green (2543 cases, 305861 assertions).

+++ src/roxy/compiler/ir/ir_optimize.cpp
@@ -694,7 +694,7 @@ CSEKey make_cse_key(IRInst* inst) {
             // share identity.
             k.a = static_cast<u32>(reinterpret_cast<uintptr_t>(
                 inst->const_data.string_val.data()));
-            k.b = static_cast<u32>(inst->const_data.string_val.size());
+            k.b = inst->const_data.string_val.size();
             break;
         case IROp::ConstNull:
             // (op, result_type) alone is sufficient — null of a given type
